Change onboarding welcome feature to theme2
