+++ feature/onboarding/welcome/src/debug/kotlin/app/k9mail/feature/onboarding/welcome/ui/WelcomeContentPreview.kt
@@ -1,0 +1,16 @@
+package app.k9mail.feature.onboarding.welcome.ui
+
+import androidx.compose.runtime.Composable
+import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
+import app.k9mail.core.ui.compose.designsystem.PreviewWithTheme
+
+@Composable
+@PreviewDevices
+internal fun WelcomeContentPreview() {
+    PreviewWithTheme {
+        WelcomeContent(
+            onStartClick = {},
+            onImportClick = {},
+        )
+    }
+}
