+++ feature/onboarding/welcome/src/main/kotlin/app/k9mail/feature/onboarding/welcome/ui/WelcomeScreen.kt
@@ -1,8 +1,6 @@
 package app.k9mail.feature.onboarding.welcome.ui
 
 import androidx.compose.runtime.Composable
-import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
-import app.k9mail.core.ui.compose.theme.ThunderbirdTheme
 
 @Composable
 fun WelcomeScreen(
@@ -13,15 +11,4 @@ fun WelcomeScreen(
         onStartClick = onStartClick,
         onImportClick = onImportClick,
     )
-}
-
-@Composable
-@PreviewDevices
-internal fun WelcomeScreenThunderbirdPreview() {
-    ThunderbirdTheme {
-        WelcomeScreen(
-            onStartClick = {},
-            onImportClick = {},
-        )
-    }
 }
